requirements.txt
flask

pygithub

pyyaml